+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
-## [0.1.0] - 2022-02-17
+## [0.0.1] - 2022-02-17
 
 - Initial release
+- Initial working version of `GithubOrgManager::Manager`
+  - Support for cloning all repos in org
+  - Support for updating all repos in org